Move MinefieldWidget and implement mode-dependent stuff in 'impl' classes
