+++ minegauler/frontend/main_window.py
@@ -215,10 +215,7 @@ class MinegaulerGUI(
         self._populate_menubars()
         self._menubar.setFixedHeight(self._menubar.sizeHint().height())
         self._panel_widget = panel.PanelWidget(self, self._state)
-        # TODO: Check which minefield type to use.
-        self._mf_widget = minefield.regular.MinefieldWidget(
-            self, self._ctrlr, self._state
-        )
+        self._mf_widget = minefield.MinefieldWidget(self, self._ctrlr, self._state)
         self.set_panel_widget(self._panel_widget)
         self.set_body_widget(self._mf_widget)
         self._name_entry_widget = _NameEntryBar(self, self._state.name)
