+++ tests/mut/frontend/minefield_test.py
@@ -15,7 +15,7 @@ from pytestqt.qtbot import QtBot
 
 from minegauler.core import api
 from minegauler.frontend import state
-from minegauler.frontend.minefield._base import _RAISED_CELL, _SUNKEN_CELL
+from minegauler.frontend.minefield._base import RAISED_CELL, SUNKEN_CELL
 from minegauler.frontend.minefield.regular import MinefieldWidget
 
 from . import utils
@@ -268,10 +268,10 @@ class TestMinefieldWidget:
         self._mouse_down_pos = pos
 
     def assert_cell_sank(self, coord):
-        self._mf_widget._set_cell_image.assert_any_call(coord, _SUNKEN_CELL)
+        self._mf_widget._set_cell_image.assert_any_call(coord, SUNKEN_CELL)
 
     def assert_cell_rose(self, coord):
-        self._mf_widget._set_cell_image.assert_any_call(coord, _RAISED_CELL)
+        self._mf_widget._set_cell_image.assert_any_call(coord, RAISED_CELL)
 
     def assert_num_cells_changed(self, num):
         """
